// src/types/enums/task.ts
export enum E_TASK_PRIORITY {
  LOW = "low",
  MEDIUM = "medium",
  HIGH = "high",
}

export enum E_TASK_STATUS {
  PENDING = "pending",
  IN_PROGRESS = "in-progress",
  COMPLETED = "completed",
}

export enum E_TASK_FIELD_NAMES {
  TITLE = "title",
  DESCRIPTION = "description",
  STATUS = "status",
  PRIORITY = "priority",
  DATE = "date",
}